(all): selectableListMixin
