+++ packages/ui/src/components/vuestic-mixins/SelectableList/SelectableListMixin.spec.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { shallowMount } from '@vue/test-utils'
 import { SelectableListMixin } from './SelectableListMixin'
 import { testIsContextableComponent } from '../../context-test/context-provide/testIsContextableComponent'
